Manual Configuration, Stack<T>, Puzzle<T,U>, PuzzleProcessor<T>, Entry points for these in Elysium::Application::Menu.
# Requirement Goals:
- Output 15-puzzle to screen.
- Correct output format.
- Output a valid puzzle.
- Output multiple 15-puzzles.
- Output multiple valid 15-puzzles.
- Prompt for user input.
- Allow user input.
- Allow program ot execute again from the interface (loop).

# Template Data Structures
The templated stack is used to store the configured puzzles, in a wrapper object (PuzzleProcessor), this should allow for fast data serialization.

# Template Puzzle.h
The puzzle wrapper was implemented as a template so that it would be possibly to extend this project for an NxN puzzle in the future, in a more convinient manner, also dynamic types can be stored in it, which may allow for easier custom extensions in the future.

# Readme
- Updated readme with build versions + status' + test status' + references.