Implemented a non-optimized algorithm to count the number of consecutive partials in a puzzle.
There is a way to calculate the total number of continuous rows/columns in the 4x4 configuration, for this I needed to implement an algorithm to get the total number of values that are continuous of size (n) in a puzzle configuration, no matter the order. The rest of the implementation will be included in the next commit.