Brute force functional for anything smaller than a 4x4. Added Threads + benchmark timer.
# Brute Force
There are a few problems with getting the brute force to work for the 4x4, it will be doable but definetely not in time for the deadline on friday morning, as I plan to submit this work tomorrow.
- I have had an idea to introduce a compile time optimization which will have to perform the brute force for the indices for each NxN matrix once, and then allow for the mapping through that to produce all the configurations of the matrix being solved, but there must be a faster solution out there, I will continue to search for it after the deadline, but for now I will probably clean this branch and then merge it with the master for the hand-in.

# Thread
I have added a thread class that is inherited by the application, which can be extended with more time, as linked in the file itself, to allow for multiple instances of this project to run simultaneously.

# Timer
I have added a timer class to benchmark the brute force solution, I utilise the chrono header library and scoping for this functionality.